+++ src/design-system/vue/components/index.ts
@@ -1,3 +1,4 @@
 export { default as DsButton } from './DsButton.vue'
 export { default as DsCard } from './DsCard.vue'
 export { default as DsInput } from './DsInput.vue'
+export { default as DsSelect } from './DsSelect.vue'